requirements.txt
Flask==0.12
Flask-Babel==0.11.2
Flask-Cors==3.0.2
Flask-Login==0.4.0
Flask-Mail==0.9.1
Flask-OpenID==1.2.5
Flask-SQLAlchemy==2.2
Flask-WhooshAlchemy==0.56
Flask-WTF==0.14.2
buttercms-python==0.7

gunicorn==19.7.1
Jinja2==2.9.4

pep8==1.7.0
python-dateutil==2.6.0

redis==2.10.5
requests==2.12.4
requests-oauthlib==0.8.0


SQLAlchemy==1.1.5
sqlalchemy-migrate==0.11.0
sqlparse==0.2.3

virtualenv==15.1.0

Werkzeug==0.11.15

WTForms==2.1

dj-database-url==0.4.1
psycopg2==2.6.2
whitenoise==2.0.6

